requirements.txt
Django>=4.2,<5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0

# Production dependencies (also work in development)
gunicorn>=21.0.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
